+++ src/main/java/com/uppy/simulations/driver_app/new_trip_received_from_client/controller/NewTripReceivedFromClientController.java
@@ -32,7 +32,7 @@ public class NewTripReceivedFromClientController implements Serializable {
                 status);
 
         int tripId = random.nextInt(Integer.MAX_VALUE);
-        double price = random.nextDouble();
+        int price = random.nextInt();
         return new TripReceivedFromClientDTO(tripId, price, "10 km",
                 "20 min", LocalDate.now().toEpochDay(), origin, destination, "Av. General Santos c/ Eligio Ayala",
                 "Feliz de Azara c/ Perú", client, 1, "Efectivo");
